feat: get, create, delete 유틸함수 생성
