+++ src/components/todo/InputTodo.tsx
@@ -2,9 +2,9 @@ import { SyntheticEvent, useCallback, useEffect, useState } from 'react';
 import './Todo.css';
 import { FaSpinner } from 'react-icons/fa';
 import useFocus from '../../hooks/useFocus';
-import { createTodo } from '../../api/todo';
 import { SetStateType, TodoDataType } from '../../types/types';
 import HandleItemButton from './HandleItemButton';
+import { handleCreateTodos } from '../../utils/todos';
 
 type Props = {
   setTodos: SetStateType<TodoDataType[]>;
@@ -19,36 +19,19 @@ const InputTodo = ({ setTodos }: Props) => {
     setFocus();
   }, [setFocus]);
 
-  const handleSubmit = useCallback(
+  const handleSubmitForm = useCallback(
     async (e: SyntheticEvent<HTMLFormElement>) => {
-      try {
-        e.preventDefault();
-        setIsLoading(true);
-
-        const trimmed = inputText.trim();
-        if (!trimmed) {
-          return alert('Please write something');
-        }
-
-        const newItem = { title: trimmed };
-        const { data } = await createTodo(newItem);
-
-        if (data) {
-          return setTodos(prev => [...prev, data]);
-        }
-      } catch (error) {
-        console.error(error);
-        alert('Something went wrong.');
-      } finally {
-        setInputText('');
-        setIsLoading(false);
-      }
+      e.preventDefault();
+      setIsLoading(true);
+      await handleCreateTodos(inputText, setTodos);
+      setIsLoading(false);
+      setInputText('');
     },
     [inputText, setTodos]
   );
 
   return (
-    <form className="form-container" onSubmit={handleSubmit}>
+    <form className="form-container" onSubmit={handleSubmitForm}>
       <input
         className="input-text"
         placeholder="Add new todo..."
